pyproject.toml
[tool.poetry]
name = "book"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.1"
django-environ = "^0.11.2"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.0"
django-ninja = "^1.1.0"
argon2-cffi-bindings = "^21.2.0"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
